+++ src/main/java/br/com/infocedro/clipper/curation/CurationCaseRepository.java
@@ -1,0 +1,6 @@
+package br.com.infocedro.clipper.curation;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+interface CurationCaseRepository extends JpaRepository<CurationCase, Long> {
+}